requirements.txt
Flask==3.1.2
pymongo==4.15.1
Werkzeug==3.1.3
python-dotenv==1.0.0
gunicorn==21.2.0
